+++ app/src/main/java/com/sungil/runningproejct_mvvm/repository/mainRepositoryImpl/MainRepoModule.kt
@@ -3,7 +3,8 @@ package com.sungil.runningproejct_mvvm.repository.mainRepositoryImpl
 import com.sungil.runningproejct_mvvm.appDatabase.RunningDao
 import com.sungil.runningproejct_mvvm.appDatabase.UserInfoDao
 import com.sungil.runningproejct_mvvm.repository.MainRepository
-import com.sungil.runningproejct_mvvm.useCase.UseCase
+import com.sungil.runningproejct_mvvm.useCase.GetUnFollowerUseCase
+import com.sungil.runningproejct_mvvm.useCase.GetFollowerUseCase
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -19,9 +20,10 @@ object MainRepoModule {
     fun providedMainRepository(
         runningDao : RunningDao,
         userDao : UserInfoDao,
-        useCase : UseCase
+        getFollowerUseCase : GetFollowerUseCase,
+        unFollowerUseCase : GetUnFollowerUseCase
     ): MainRepository  {
-        return MainRepositoryImpl(runningDao , userDao ,useCase )
+        return MainRepositoryImpl(runningDao , userDao ,getFollowerUseCase , unFollowerUseCase )
     }
